LayoutTests/http/tests/security/contentSecurityPolicy/source-list-parsing-paths-01-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/RESOURCES/script.js because it does not appear in the script-src directive of the Content Security Policy.
Resources should be rejected unless they match a whitelisted path.



--------
Frame: '<!--frame1-->'
--------
PASS

--------
Frame: '<!--frame2-->'
--------
PASS

--------
Frame: '<!--frame3-->'
--------
PASS

--------
Frame: '<!--frame4-->'
--------
PASS

--------
Frame: '<!--frame5-->'
--------
PASS

--------
Frame: '<!--frame6-->'
--------
PASS

--------
Frame: '<!--frame7-->'
--------
PASS

--------
Frame: '<!--frame8-->'
--------
PASS
